backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
pydantic==2.10.4
google-generativeai==0.8.3
PyMuPDF==1.25.1
numpy==1.26.4
faiss-cpu==1.9.0
groq==0.11.0
sqlalchemy==2.0.36
passlib[argon2]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.12
pydantic[email]==2.10.4
